implement review suggestions, refactor code
