+++ sessions/react-global-state/demo-end/pages/components/Layout.js
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import Link from "next/link";
 
 import Header from "./Header";
 import Footer from "./Footer";
@@ -11,6 +10,7 @@ const StyledHeader = styled.header`
   width: 100%;
   padding: 1rem;
   align-items: center;
+  justify-content: center;
   background: var(--primary-color);
   color: #eee;
 `;
@@ -22,28 +22,25 @@ const StyledFooter = styled.footer`
   width: 100%;
   padding: 1rem;
   align-items: center;
+  justify-content: center;
   background: var(--primary-color);
   color: #eee;
 `;
 
 const Container = styled.div`
   display: flex;
+  margin: auto;
+  max-width: 80%;
   flex-direction: column;
-  padding: 4rem 1rem;
+  padding: 5rem 1rem;
+  gap: 1rem;
 `;
 
-const StyledLink = styled(Link)`
-  text-decoration: none;
-  color: #fff;
-`;
-
-export default function AppLayout({ countSum, dragonCount, children }) {
+export default function Layout({ countSum, dragonCount, children }) {
   return (
     <>
       <StyledHeader>
-        <StyledLink href="/">
-          <Header countSum={countSum} />
-        </StyledLink>
+        <Header countSum={countSum} />
       </StyledHeader>
       <Container>{children}</Container>
       <StyledFooter>
