+++ sessions/react-global-state/demo-end/pages/index.js
@@ -1,10 +1,14 @@
 import styled from "styled-components";
 import Link from "next/link";
 
-const Flex = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
+import CounterStats from "./components/CounterStats";
+
+const NavWrapper = styled.nav`
+  padding-inline: 1rem;
+`;
+
+const MainWrapper = styled.main`
+  padding-inline: 1rem;
 `;
 
 const StyledLink = styled(Link)`
@@ -14,23 +18,18 @@ const StyledLink = styled(Link)`
 
 export default function HomePage({ countSum, dragonCount, countAverage }) {
   return (
-    <Flex>
-      <div>
-        <h2>Counter Stats</h2>
-        <p>
-          Our counters currently have an average count of{" "}
-          <strong>{countAverage}</strong>, counting <strong>{countSum}</strong>{" "}
-          things ... of which <strong>{dragonCount}</strong> are dragons.
-        </p>
-        <h2>Definitions</h2>
-        <dl>
-          <dt>Counter</dt>
-          <dd>A counter is a thing that can count up or down.</dd>
-          <dd>A counter is also a piece of furniture.</dd>
-        </dl>
-      </div>
+    <>
+      <MainWrapper>
+        <CounterStats
+          countSum={countSum}
+          dragonCount={dragonCount}
+          countAverage={countAverage}
+        />
+      </MainWrapper>
 
-      <StyledLink href="/overview">To the counters ➡︎</StyledLink>
-    </Flex>
+      <NavWrapper>
+        <StyledLink href="/counters">To the counters →</StyledLink>
+      </NavWrapper>
+    </>
   );
 }
